+++ CHANGELOG.md
@@ -1,5 +1,32 @@
 # Changelog
 
+## [0.4.1](https://github.com/optiplex331/Halligalli-BossYang/compare/v0.4.0...v0.4.1) (2026-06-10)
+
+
+### Bug Fixes
+
+* **deploy:** release backend-only Azure image ([87c08e3](https://github.com/optiplex331/Halligalli-BossYang/commit/87c08e3c186fd4648e5908ac20c57ff4dd239e28))
+
+## [0.4.0](https://github.com/optiplex331/Halligalli-BossYang/compare/v0.3.0...v0.4.0) (2026-06-04)
+
+
+### Features
+
+* **deploy:** switch production templates to AWS ([#30](https://github.com/optiplex331/Halligalli-BossYang/issues/30)) ([91605af](https://github.com/optiplex331/Halligalli-BossYang/commit/91605afda7377dc5d6186bdadd1df94d6b2e4978))
+
+## [0.3.0](https://github.com/optiplex331/Halligalli-BossYang/compare/v0.2.0...v0.3.0) (2026-06-03)
+
+
+### Features
+
+* **aws-staging:** scaffold portfolio environment ([4155db5](https://github.com/optiplex331/Halligalli-BossYang/commit/4155db5d7c0636f1f785c3042bfb5dd426d71c01))
+* **ci:** harden aws staging activation ([1db0d31](https://github.com/optiplex331/Halligalli-BossYang/commit/1db0d311bbb313a898679a0e121f36a32586ddfa))
+
+
+### Bug Fixes
+
+* **app:** sync Vite entrypoint and release docs ([539ff83](https://github.com/optiplex331/Halligalli-BossYang/commit/539ff83077d32dc31c9e31667e7c7eb850d4d3d0))
+
 ## [0.2.0](https://github.com/optiplex331/Halligalli-BossYang/compare/v0.1.0...v0.2.0) (2026-05-23)